Abstract Calendar rendering into dedicated component
